+++ src/navigators/StackNavigator.tsx
@@ -41,7 +41,7 @@ function StackNavigator() {
   return (
     <Stack.Navigator>
       {/*로그인 페이지*/}
-      {/*<Stack.Screen
+      <Stack.Screen
         name={stackScreens.LoginScreen}
         component={LoginScreen}
         options={{headerShown: false}}
@@ -55,7 +55,7 @@ function StackNavigator() {
         name={stackScreens.SignupCompleteScreen}
         component={SignupCompleteScreen}
         options={{headerShown: false}}
-      />*/}
+      />
       {/*BottomTab 3개 페이지*/}
       <Stack.Screen
         name={stackScreens.BottomTabScreens}
